Conflicts:
	extra/language-packs/po/fr.po

+++ main/dns/src/EBox/DNS.pm
@@ -22,7 +22,6 @@ use base qw( EBox::Module::Service
              EBox::SysInfo::Observer
              EBox::NetworkObserver );
 
-use EBox::Objects;
 use EBox::Gettext;
 use EBox::Config;
 use EBox::Service;
@@ -42,7 +41,7 @@ use EBox::Exceptions::UnwillingToPerform;
 use EBox::Exceptions::DataNotFound;
 use EBox::Exceptions::MissingArgument;
 
-use TryCatch::Lite;
+use TryCatch;
 use File::Temp;
 use File::Slurp;
 use Fcntl qw(:seek);
@@ -345,6 +344,68 @@ sub hostIpAddresses
     return $array;
 }
 
+# Method: getAddresses
+#
+#   Given a domain name, it returns an array ref of addresses that
+#   it contains.
+#
+# Parameters:
+#
+#   domain - String the domain's name
+#
+# Returns:
+#
+#   array ref - list of IP addresses
+#
+sub getAddresses
+{
+    my ($self, $domain) = @_;
+
+    unless (defined $domain) {
+        throw EBox::Exceptions::MissingArgument('domain');
+    }
+
+    my $domainRow = $self->model('DomainTable')->findRow(domain => $domain);
+    unless (defined $domainRow) {
+        throw EBox::Exceptions::DataNotFound(data  => __('domain'),
+                                             value => $domain);
+    }
+
+    return $self->_domainIpAddresses($domainRow->subModel('ipAddresses'));
+}
+
+# Method: allAddressesInUse
+#
+#   Returns a hash ref with all IPs of all domains
+#
+# Returns:
+#
+#  array ref with this structure data:
+#      ip - domain/host where used
+#
+sub allAddressesInUse
+{
+    my ($self) = @_;
+
+    my %ips;
+
+    foreach my $d (@{$self->domains()}) {
+        my $domain = $d->{name};
+        foreach my $h (@{$self->getHostnames($domain)}) {
+            foreach my $ip (@{$h->{ip}}) {
+                next unless $ip;
+                $ips{$ip} = $h->{name} . '.' . $domain;
+            }
+        }
+        foreach my $ip (@{$self->getAddresses($domain)}) {
+            next unless $ip;
+            $ips{$ip} = $domain;
+        }
+    }
+
+    return \%ips;
+}
+
 # Method: getServices
 #
 #   Given a domain name, it returns an array ref of SRV records that
@@ -564,7 +625,7 @@ sub initialSetup
 
     # Create default rules and services only if installing the first time
     unless ($version) {
-        my $services = EBox::Global->modInstance('services');
+        my $services = EBox::Global->modInstance('network');
 
         my $serviceName = 'dns';
         unless ($services->serviceExists(name => $serviceName)) {
@@ -612,22 +673,7 @@ sub _daemons
 {
     return [
         {
-            'name' => 'bind9',
-            'type' => 'init.d'
-        }
-    ];
-}
-
-# Method: _daemonsToDisable
-#
-#  Overrides <EBox::Module::Service::_daemonsToDisable>
-#
-sub _daemonsToDisable
-{
-    return [
-        {
-            'name' => 'bind9',
-            'type' => 'init.d'
+            'name' => 'bind9'
         }
     ];
 }
@@ -640,9 +686,6 @@ sub _preSetConf
     my ($self) = @_;
 
     my $runResolvConf = 1;
-    if ($self->global->modExists('samba')) {
-        my $usersModule = $self->global->modInstance('samba');
-    }
     my $array = [];
     push (@{$array}, runResolvConf => $runResolvConf);
     $self->writeConfFile(BIND9DEFAULTFILE, 'dns/bind9.mas', $array,
@@ -918,18 +961,38 @@ sub _postServiceHook
 {
     my ($self, $enabled) = @_;
 
-    if ($enabled) {
+    my $samba = $self->global->modInstance('samba');
+    # TODO: separate regular nsupdateCmds with -l from dlz ones with -g
+    if ($enabled and defined($samba)) {
+        # Wait max of 30 seconds until named is listening
         my $nTry = 0;
         do {
             sleep(1);
-        } while ( $nTry < 5 and (not $self->_isNamedListening()));
-        if ( $nTry < 5 ) {
+        } while (($nTry < 30) and (not $self->_isPortListening(53)));
+        # Do nothing if Kerberos is not listening
+        if (($nTry < 5) and $self->_isPortListening(88)) {
+            my $dnsFile = new File::Temp(TEMPLATE => 'resolvXXXXXX', DIR => EBox::Config::tmp());
+            EBox::Sudo::root("cp /etc/resolvconf/interface-order $dnsFile",
+                             'echo zentyaldns.temp > /etc/resolvconf/interface-order',
+                             "echo 'nameserver 127.0.0.1' | resolvconf -a zentyaldns.temp");
+            my $keytabPath = EBox::Samba::SAMBA_DNS_KEYTAB();
+            my $hostname = $self->global()->modInstance('sysinfo')->hostName();
+            my $netbiosName = $samba->model('DomainSettings')->value('netbiosName');
+            foreach my $name ($hostname, $netbiosName, uc($netbiosName)) {
+                EBox::Sudo::silentRoot("samba-tool user list | grep ^dns-$name");
+                if ($? == 0) {
+                    EBox::Sudo::root("kinit -k -t $keytabPath dns-$name");
+                    last;
+                }
+            }
             foreach my $cmd (@{$self->{nsupdateCmds}}) {
                 EBox::Sudo::root($cmd);
                 my ($filename) = $cmd =~ m:\s(.*?)$:;
                 # Remove the temporary file
                 unlink ($filename) if -f $filename;
             }
+            EBox::Sudo::root("cp $dnsFile /etc/resolvconf/interface-order",
+                             'resolvconf -d zentyaldns.temp');
             delete $self->{nsupdateCmds};
         }
     }
@@ -1420,7 +1483,7 @@ sub _updateDynReverseZone
         unshift(@file, "zone $zone");
         push(@file, "send");
         untie(@file);
-        $self->_launchNSupdate($fh);
+        $self->_launchNSupdate($fh, 1);
     }
 }
 
@@ -1522,23 +1585,24 @@ sub _removeDeletedRR
 #
 sub _launchNSupdate
 {
-    my ($self, $fh) = @_;
+    my ($self, $fh, $reverse) = @_;
 
-    my $cmd = NS_UPDATE_CMD . ' -l -t 10 ' . $fh->filename();
+    my $auth = $reverse ? '-l' : '-g';
+    my $cmd = NS_UPDATE_CMD . " $auth -t 10 " . $fh->filename();
     $self->{nsupdateCmds} = [] unless exists $self->{nsupdateCmds};
     push (@{$self->{nsupdateCmds}}, $cmd);
     $fh->unlink_on_destroy(0);
 }
 
-# Check if named is listening
-sub _isNamedListening
+# Check if port is listening
+sub _isPortListening
 {
-    my ($self) = @_;
+    my ($self, $port) = @_;
 
     my $sock = new IO::Socket::INET(PeerAddr => '127.0.0.1',
-                                    PeerPort => 53,
+                                    PeerPort => $port,
                                     Proto    => 'tcp');
-    if ( $sock ) {
+    if ($sock) {
         close($sock);
         return 1;
     } else {
@@ -1970,6 +2034,18 @@ sub hostDomainChangedDone
                 last;
             }
         }
+    }
+}
+
+sub checkDuplicatedIP
+{
+    my ($self, $ip) = @_;
+
+    return if EBox::Config::boolean('allow_duplicated_ips');
+
+    my $domain = $self->allAddressesInUse()->{$ip};
+    if ($domain) {
+        throw EBox::Exceptions::DataInUse(__x("The IP '{a}' already assigned in '{d}'", a => $ip, d => $domain));
     }
 }
